UI Update 10/15
-Integrated the Carbon dial, which will serve as our measure of carbon for the gameplay.
-Reorganized the UI a bit, still needs work
-Adjusted the contracts so they now properly close, open, and are a bit neater.
NOTE: I need to look through the contract UI code to clean it up. It has a lot of pieces that I don't think are needed with the current direction of the game. I have preserved the old UI, however, so that if we need any of the old material we can still use it.